bitcoin_source_producer/requirements.txt
quixstreams~=2.8.0
websocket-client==1.8.0